Refact community actions to action-creators folder convention #362
