+++ client/community/action-creators/index.js
@@ -1,1 +1,6 @@
 export { default as asyncCreate } from './async-create'
+export { default as asyncDownloadActivists } from './async-download-activists'
+export { default as asyncEdit } from './async-edit'
+export { default as asyncFetch } from './async-fetch'
+export { default as select } from './select'
+export { default as unset } from './unset'
