+++ client/community/action-creators/select.js
@@ -1,0 +1,7 @@
+import * as t from '~community/action-types'
+
+const select = id => dispatch => {
+  return dispatch({ type: t.SELECT, id })
+}
+
+export default select
