+++ client/community/action-creators/unset.js
@@ -1,0 +1,5 @@
+import * as t from '~community/action-types'
+
+const unset = dispatch => dispatch({ type: t.UNSET })
+
+export default unset
