+++ client/community/index.js
@@ -2,12 +2,10 @@ import createRoutes from './createRoutes'
 import reducers from './reducers'
 import * as paths from './paths'
 import * as selectors from './selectors'
-import { fetch } from './actions'
 
 export default {
   createRoutes,
   reducers,
   selectors,
-  fetch,
   paths
 }
