+++ routes/logged-in/background/community-list/page.connected.js
@@ -1,7 +1,7 @@
 import { provideHooks } from 'redial'
 import { connect } from 'react-redux'
 
-import * as CommunityActions from '~community/actions'
+import * as CommunityActions from '~community/action-creators'
 import * as CommunitySelectors from '~community/selectors'
 
 import Page from './page'
@@ -12,7 +12,7 @@ const redial = {
     const promises = []
 
     !CommunitySelectors.isLoaded(state) && promises.push(
-      dispatch(CommunityActions.fetch())
+      dispatch(CommunityActions.asyncFetch())
     )
     return promises
   }
