+++ routes/logged-in/sidebar/community-settings-info/page.connected.js
@@ -2,7 +2,7 @@ import { provideHooks } from 'redial'
 import { connect } from 'react-redux'
 import { reduxForm } from 'redux-form'
 
-import * as CommunityActions from '~community/actions'
+import * as CommunityActions from '~community/action-creators'
 import * as CommunitySelectors from '~community/selectors'
 
 import Page from './page'
@@ -13,7 +13,7 @@ const redial = {
     const promises = []
 
     !CommunitySelectors.isLoaded(state) && promises.push(
-      dispatch(CommunityActions.fetch())
+      dispatch(CommunityActions.asyncFetch())
     )
     !CommunitySelectors.getCurrentId(state) && promises.push(
       dispatch(CommunityActions.select(1))
@@ -31,8 +31,8 @@ const mapStateToProps = state => {
 }
 
 const mapDispatchToProps = {
-  submit: CommunityActions.edit,
-  downloadActivists: CommunityActions.downloadActivists
+  submit: CommunityActions.asyncEdit,
+  downloadActivists: CommunityActions.asyncDownloadActivists
 }
 
 const fields = ['id', 'image', 'name', 'city', 'description']
