+++ routes/logged-in/sidebar/community-settings-mailchimp/page.connected.js
@@ -2,7 +2,7 @@ import { provideHooks } from 'redial'
 import { connect } from 'react-redux'
 import { reduxForm } from 'redux-form'
 
-import * as CommunityActions from '~community/actions'
+import * as CommunityActions from '~community/action-creators'
 import * as CommunitySelectors from '~community/selectors'
 
 import Page from './page'
@@ -13,7 +13,7 @@ const redial = {
     const promises = []
 
     !CommunitySelectors.isLoaded(state) && promises.push(
-      dispatch(CommunityActions.fetch())
+      dispatch(CommunityActions.asyncFetch())
     )
     !CommunitySelectors.getCurrentId(state) && promises.push(
       dispatch(CommunityActions.select(1))
@@ -27,7 +27,7 @@ const mapStateToProps = state => ({
 })
 
 const mapDispatchToProps = {
-  submit: CommunityActions.edit
+  submit: CommunityActions.asyncEdit
 }
 
 export default provideHooks(redial)(
